fix: run TxFinalizationSection(...) before worldUpdater gets discarded
The WorldUpdater of the frame gets discarded in case of an exceptional
halt. Besu does this BEFORE it traceContextExit. This fix does creates
the

  TxFinalizationSection

before the wiping of the WorldUpdater so as to extract the "unupdated"
balances.

Note. We will have to apply a similar fix to other operations that make
snapshots at the boundaries of context changes: CALL's, CREATE's,
RETURN's and SELFDESTRUCT's to name a few, and more generally those
opcodes that are REVERT sensitive.